More plugin loading shenanigans
1. Basically I think I have a neat solution for plugin configuration

On VimEnter au triggers a command Vimenter
All plugins that need to be deferred are loaded with {'on': 'Vimenter'}
So vim loads, plugins are deferred, but once Vimenter autocmd triggers, the plugins are loaded 'lazily' and that triggers the user autocommand to source the configuration script.

So plugin config can remain in one file that's guaranteed to be loaded.